% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluon-utils.R
\name{ag_tabular_module}
\alias{ag_tabular_module}
\title{AutoGluon Tabular Function}
\usage{
ag_tabular_module(target = "target_variable")
}
\arguments{
\item{target}{a variable to "target" for within the training data}
}
\value{
a list of the initial fit (for other processes) and the predictions
}
\description{
AutoGluon Tabular Function
}
