% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluon-utils.R
\name{autogluon_tabular_fit_info}
\alias{autogluon_tabular_fit_info}
\title{AutoGluon Tabular Fitted Model In-Sample Information}
\usage{
autogluon_tabular_fit_info(object)
}
\arguments{
\item{object}{a fitted autogluons model of class (autogluon.tabular.predictor.predictor.TabularPredictor)}
}
\value{
a list of in-sample metrics with model and data information
}
\description{
AutoGluon Tabular Fitted Model In-Sample Information
}
