% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_quark.R
\name{chat_quark}
\alias{chat_quark}
\title{chat with quark}
\usage{
chat_quark(
  provider = "ollama",
  model = "mistral-nemo",
  prompt =
    "please use 'mtcars_train' as training data and 'mtcars_test' as testing data and run the tabular module. I want to predict mpg.",
  flavor = "concise"
)
}
\arguments{
\item{provider}{an LLM provider}

\item{model}{a model from given provider that supports \code{tools}}

\item{prompt}{a user prompt to tell the LLM to train and predict with autogluons}

\item{flavor}{a quark "style", either "concise", "detailed", or "refined"}
}
\value{
an LLM chat interface which can return a fitted model and predictions
}
\description{
chat with quark
}
