% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python-utils.R
\name{create_quark_env}
\alias{create_quark_env}
\title{Create Virtual Environment Wrapper}
\usage{
create_quark_env(envname = "quark", ...)
}
\arguments{
\item{envname}{virtual environment to create}

\item{...}{additional passes for \code{create_virtualenv}}
}
\value{
creation of virtual environment
}
\description{
Create Virtual Environment Wrapper
}
