% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python-utils.R
\name{install_quark_dependencies}
\alias{install_quark_dependencies}
\title{Install AutoGluons}
\usage{
install_quark_dependencies(envname = "quark", method = "auto", ...)
}
\arguments{
\item{envname}{virtual environment name}

\item{method}{method defaults to "auto"}

\item{...}{additional passes for \code{py_install}}
}
\value{
installs autogluon to "quark" by default
}
\description{
Install AutoGluons
}
