% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluon-utils.R
\name{load_trained_model_and_predict}
\alias{load_trained_model_and_predict}
\title{most recently trained gluons model and test preds}
\usage{
load_trained_model_and_predict()
}
\value{
list of python.builtin.object and predictions
}
\description{
most recently trained gluons model and test preds
}
