% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quark-package.R
\docType{package}
\name{quark-package}
\alias{quark}
\alias{quark-package}
\title{quark: quark, a gluon agent}
\description{
quark is an autogluons agent for agentic analytics. it's a simple reticulate binding with ellmer tooling. an experimentation of fitting & predicting information as a service.
}
\author{
\strong{Maintainer}: frank hull \email{frankiethull@gmail.com}

}
\keyword{internal}
