% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluon-utils.R
\name{save_testing_data_csv}
\alias{save_testing_data_csv}
\title{Saves a testing data as csv}
\usage{
save_testing_data_csv(testing_data = "mtcars")
}
\arguments{
\item{testing_data}{The testing dataframe}
}
\value{
none
}
\description{
Saves a testing data as csv
}
