% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-utils.R
\name{save_testing_data_csv_tool}
\alias{save_testing_data_csv_tool}
\title{save off data as testing.csv}
\usage{
save_testing_data_csv_tool(testing_data = "mtcars")
}
\description{
save off data as testing.csv
}
