% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluon-utils.R
\name{save_training_data_csv}
\alias{save_training_data_csv}
\title{Saves a training data as csv}
\usage{
save_training_data_csv(training_data = "mtcars")
}
\arguments{
\item{training_data}{The training dataframe}
}
\value{
none
}
\description{
Saves a training data as csv
}
