% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-utils.R
\name{save_training_data_csv_tool}
\alias{save_training_data_csv_tool}
\title{save off data as training.csv}
\usage{
save_training_data_csv_tool(training_data = "mtcars")
}
\description{
save off data as training.csv
}
