% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluon-utils.R
\name{summarize_and_save_testing_data_csv}
\alias{summarize_and_save_testing_data_csv}
\title{summarizes and saves testing data as csv}
\usage{
summarize_and_save_testing_data_csv(testing_data = "mtcars")
}
\arguments{
\item{testing_data}{The testing dataframe}
}
\value{
saves dataset and returns a summary for context
}
\description{
summarizes and saves testing data as csv
}
