% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-utils.R
\name{summarize_and_save_testing_data_csv_tool}
\alias{summarize_and_save_testing_data_csv_tool}
\title{summarize and save off data as testing.csv}
\usage{
summarize_and_save_testing_data_csv_tool(testing_data = "mtcars")
}
\description{
summarize and save off data as testing.csv
}
