% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluon-utils.R
\name{summarize_and_save_training_data_csv}
\alias{summarize_and_save_training_data_csv}
\title{summarizes and saves a training data as csv}
\usage{
summarize_and_save_training_data_csv(training_data = "mtcars")
}
\arguments{
\item{training_data}{The training dataframe}
}
\value{
saves dataset and returns a summary for context
}
\description{
summarizes and saves a training data as csv
}
