% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gluon-utils.R
\name{summarize_split_and_save_data_csv}
\alias{summarize_split_and_save_data_csv}
\title{summarizes, splits, and saves datasets}
\usage{
summarize_split_and_save_data_csv(all_data = "mtcars")
}
\arguments{
\item{all_data}{}
}
\value{
splits data into training/testing, saves datasets, and returns a data summary for context
}
\description{
if the user shares an entire data, instead of train/test,
this tool will split the data for the user and stage testing and training for them
}
