% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool-utils.R
\name{summarize_split_and_save_data_csv_tool}
\alias{summarize_split_and_save_data_csv_tool}
\title{summarize all data, split into testing/training, save off data as training.csv & testing.csv}
\usage{
summarize_split_and_save_data_csv_tool(all_data = "mtcars")
}
\description{
summarize all data, split into testing/training, save off data as training.csv & testing.csv
}
