% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python-utils.R
\name{use_quark_env}
\alias{use_quark_env}
\title{Use Virtual Environment Wrapper}
\usage{
use_quark_env(envname = "quark", ...)
}
\arguments{
\item{envname}{virtual environment to use}

\item{...}{additional passes for \code{use_virtualenv}}
}
\value{
sets environment to virtualenv
}
\description{
Use Virtual Environment Wrapper
}
